pyproject.toml
[tool.poetry]
name = "chirp"
version = "0.1.0"
description = "A bioacoustics research project."
authors = ["Chirp Team <chirp-bio@google.com>"]
license = "Apache 2.0"

[tool.poetry.dependencies]
python = ">=3.7.1,<3.11"
absl-py = "^1.0.0"
clu = "^0.0.7"
flax = "^0.5.1"
jax = "^0.3.9"
ml-collections = "^0.1.1"
optax = "^0.1.2"
tensorflow = "^2.8.0"
tensorflow-datasets = {extras = ["dev"], version = "^4.6.0"}
pandas = "==1.3.5"
etils = {extras = ["epath"], version = "^0.6.0"}
ratelimiter = "^1.2.0.post0"
SPARQLWrapper = "^2.0.0"

[tool.poetry.group.dev.dependencies]
librosa = "^0.9.1"
numpy = "^1.21.6"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
